DirectTest checks compiler options
Previously, a bad compiler option was ignored.

Add a common idiom to StreamCapture to set stdout;
the idiom may be suboptimal for testing.

Convert a partest to junit for regex, and eliminate
one check file.
