+++ src/partest/scala/tools/partest/nest/StreamCapture.scala
@@ -28,6 +28,12 @@ object StreamCapture {
     }
   }
 
+  /** Set err to out. */
+  def redirErr[A](body: => A): A = savingSystem {
+    System.setOut(System.err)
+    body
+  }
+
   def capturingOutErr[A](output: OutputStream)(body: => A): A = {
     val charset = Charset.defaultCharset()
     Using.resource(new PrintStream(output, /*autoflush=*/true, charset.name())) { printStream =>
