--- test/files/run/t6178.check
@@ -1,1 +1,0 @@
-2
