+++ test/files/run/t6178.scala
@@ -2,6 +2,6 @@ import scala.reflect.runtime.universe._
 import scala.reflect.runtime.{currentMirror => cm}
 
 object Test extends App {
-  val plus = typeOf[java.lang.String].member(TermName("$plus")).asMethod
-  println(cm.reflect("").reflectMethod(plus).apply("2"))
+  val plusMethod = typeOf[java.lang.String].member(TermName("$plus")).asMethod
+  assert(cm.reflect("").reflectMethod(plusMethod).apply("2") == "2")
 }
